+++ home_work_1/cython/v3/home_work.pyx
@@ -5,18 +5,18 @@ from libcpp.unordered_map cimport unordered_map
 cpdef target_mean(data,y_name,x_name):
     cdef long nrow = data.shape[0]
     cdef np.ndarray[double] result = np.asfortranarray(np.zeros(nrow),dtype=np.float64)
-    cdef np.ndarray[int] y = np.asfortranarray(data[y_name],dtype=np.int32)
-    cdef np.ndarray[int] x = np.asfortranarray(data[x_name],dtype=np.int32)
+    cdef np.ndarray[double] y = np.asfortranarray(data[y_name],dtype=np.int32)
+    cdef np.ndarray[double] x = np.asfortranarray(data[x_name],dtype=np.int32)
 
     target_mean_impl(result, y, x, nrow)
     return result
 
 
-cdef void target_mean_impl(double[:] result, int[:] y, int[:] x, const long nrow):
+cdef void target_mean_impl(double[:] result, double[:] y, double[:] x, const long nrow):
 #    cdef dict value_dict = dict()
 #    cdef dict count_dict = dict()
-    cdef unordered_map[int, int] value_map
-    cdef unordered_map[int, int] count_map
+    cdef unordered_map[double, double] value_map
+    cdef unordered_map[double, double] count_map
     cdef long i
     for i in range(nrow):
         if value_map[x[i]] == None:
